bot delete methods fixed

+++ project/OsEngine/OsTrader/Panels/BotPanel.cs
@@ -158,6 +158,7 @@ namespace OsEngine.OsTrader.Panels
 
         void _chartUi_Closed(object sender, EventArgs e)
         {
+            _chartUi.Closed -= _chartUi_Closed;
             _chartUi = null;
 
             if (ChartClosedEvent != null)
@@ -267,17 +268,22 @@ namespace OsEngine.OsTrader.Panels
                 for (int i = 0; _botTabs != null && i < _botTabs.Count; i++)
                 {
                     _botTabs[i].StopPaint();
-                    try
-                    {
-                        _log.StopPaint();
-                    }
-                    catch(Exception error)
-                    {
-                        LogMessageEvent(error.ToString(), LogMessageType.Error);
-                    }
-
                 }
 
+                try
+                {
+                    _log.StopPaint();
+                }
+                catch (Exception error)
+                {
+                    LogMessageEvent(error.ToString(), LogMessageType.Error);
+                }
+
+                if(_tabBotTab != null)
+                {
+                    _tabBotTab.SelectionChanged -= _tabBotTab_SelectionChanged;
+                }
+           
                 _tabBotTab = null;
                 _tabBotTab = null;
                 _hostChart = null;
@@ -385,27 +391,67 @@ namespace OsEngine.OsTrader.Panels
         {
             try
             {
-                _riskManager.Delete();
+                if (_riskManager != null)
+                {
+                    _riskManager.RiskManagerAlarmEvent -= _riskManager_RiskManagerAlarmEvent;
+                    _riskManager.Delete();
+                    _riskManager = null;
+                }
 
                 if (_botTabs != null)
                 {
                     for (int i = 0; i < _botTabs.Count; i++)
                     {
+                        _botTabs[i].StopPaint();
                         _botTabs[i].Delete();
+                        _botTabs[i].LogMessageEvent -= SendNewLogMessage;
                     }
                     _botTabs.Clear();
                     _botTabs = null;
                 }
 
-                if (File.Exists(@"Engine\" + NameStrategyUniq + @"Parametrs.txt"))
+                if(ParamGuiSettings != null)
                 {
-                    File.Delete(@"Engine\" + NameStrategyUniq + @"Parametrs.txt");
+                    ParamGuiSettings.LogMessageEvent -= SendNewLogMessage;
+                    ParamGuiSettings = null;
+
+                    if (File.Exists(@"Engine\" + NameStrategyUniq + @"Parametrs.txt"))
+                    {
+                        File.Delete(@"Engine\" + NameStrategyUniq + @"Parametrs.txt");
+                    }
                 }
 
                 if(_log != null)
                 {
                     _log.Delete();
+                    _log = null;
                 }
+
+                if(_parameters != null)
+                {
+                    for(int i = 0;i < _parameters.Count;i++)
+                    {
+                        _parameters[i].ValueChange -= Parameter_ValueChange;
+                    }
+                    _parameters.Clear();
+                    _parameters = null;
+                }
+
+                if(_tabBotTab != null)
+                {
+                    _tabBotTab.SelectionChanged -= _tabBotTab_SelectionChanged;
+                    _tabBotTab = null;
+                }
+
+                _gridChart = null;
+                _hostChart = null;
+                _hostGlass = null;
+                _hostOpenDeals = null;
+                _hostCloseDeals = null;
+                _rectangle = null;
+                _hostAlerts = null;
+                _textBoxLimitPrice = null;
+                _gridChartControlPanel = null;
 
                 if (DeleteEvent != null)
                 {
@@ -1260,7 +1306,9 @@ position => position.State != PositionStateType.OpeningFail
                 }
 
                 ActivTab.Delete();
+
                 _botTabs.Remove(ActivTab);
+
                 if (_botTabs != null && _botTabs.Count != 0)
                 {
                     ChangeActivTab(0);
